homework: split out password cache logic into its own .c/.h file
Preparation for extending it further down the line.
