+++ src/home/homework.h
@@ -6,8 +6,8 @@
 
 #include "sd-id128.h"
 
+#include "homework-password-cache.h"
 #include "loop-util.h"
-#include "strv.h"
 #include "user-record.h"
 #include "user-record-util.h"
 
@@ -42,22 +42,6 @@ typedef struct HomeSetup {
 
         char *temporary_image_path;
 } HomeSetup;
-
-typedef struct PasswordCache {
-        /* Decoding passwords from security tokens is expensive and typically requires user interaction,
-         * hence cache any we already figured out. */
-        char **pkcs11_passwords;
-        char **fido2_passwords;
-} PasswordCache;
-
-void password_cache_free(PasswordCache *cache);
-
-static inline bool password_cache_contains(const PasswordCache *cache, const char *p) {
-        if (!cache)
-                return false;
-
-        return strv_contains(cache->pkcs11_passwords, p) || strv_contains(cache->fido2_passwords, p);
-}
 
 #define HOME_SETUP_INIT                                 \
         {                                               \
